Refactoring type conversions
Removing as many unnecessary type conversions as possible, especially for DNA/protein
characters (u8 -> char, u8 -> usize).
