+++ phylo/src/sequences/mod.rs
@@ -7,29 +7,23 @@ pub enum SequenceType {
     Protein,
 }
 
-pub fn charify(chars: &str) -> Vec<u8> {
-    chars.chars().map(|c| c as u8).collect()
-}
-
-pub static AMINOACIDS_STR: &str = "ARNDCQEGHILKMFPSTWYV";
-pub static AMB_AMINOACIDS_STR: &str = "BJZX";
-
-pub static NUCLEOTIDES_STR: &str = "TCAG";
-pub static AMB_NUCLEOTIDES_STR: &str = "RYSWKMBDHVNZX";
-
-pub(crate) static GAP: u8 = b'-';
+pub static AMINOACIDS: &[u8] = b"ARNDCQEGHILKMFPSTWYV";
+pub static AMB_AMINOACIDS: &[u8] = b"BJZX";
+pub static NUCLEOTIDES: &[u8] = b"TCAG";
+pub static AMB_NUCLEOTIDES: &[u8] = b"RYSWKMBDHVNZX";
+pub static GAP: u8 = b'-';
 
 pub fn dna_alphabet() -> Alphabet {
-    let mut nucleotides = charify(NUCLEOTIDES_STR);
-    nucleotides.append(&mut (charify(AMB_NUCLEOTIDES_STR)));
+    let mut nucleotides = NUCLEOTIDES.to_vec();
+    nucleotides.append(&mut AMB_NUCLEOTIDES.to_vec());
     nucleotides.append(&mut nucleotides.clone().to_ascii_lowercase());
     nucleotides.push(GAP);
     Alphabet::new(nucleotides)
 }
 
 pub fn protein_alphabet() -> Alphabet {
-    let mut aminoacids = charify(AMINOACIDS_STR);
-    aminoacids.append(&mut (charify(AMB_AMINOACIDS_STR)));
+    let mut aminoacids = AMINOACIDS.to_vec();
+    aminoacids.append(&mut AMB_AMINOACIDS.to_vec());
     aminoacids.append(&mut aminoacids.clone().to_ascii_lowercase());
     aminoacids.push(GAP);
     Alphabet::new(aminoacids)
